+++ package_documentation.py
@@ -6,7 +6,7 @@ prefix = pathlib.Path(os.getenv("PREFIX")).resolve()
 sp_dir = pathlib.Path(os.getenv("SP_DIR")).resolve()
 pkg_name = os.getenv("PKG_NAME")
 
-man_path = pathlib.Path("build/docs/man").resolve()
+man_path = pathlib.Path("build/docs/man/waves.1").resolve()
 html_path_external = pathlib.Path("build/docs/html").resolve()
 html_path_internal = pathlib.Path("build/docs/html-internal").resolve()
 html_path = html_path_external if html_path_external.exists() else html_path_internal
@@ -34,4 +34,5 @@ for destination, source in new_paths:
         shutil.copy2(source, destination, follow_symlinks=True)
     else:
         destination.mkdir(parents=True, exist_ok=True)
-        shutil.copytree(source, destination, symlinks=False, dirs_exist_ok=True)
+        shutil.copytree(source, destination, symlinks=False, dirs_exist_ok=True,
+                        ignore=shutil.ignore_patterns(".doctrees", "*.doctree", ".buildinfo"))